seperates menu into its own module
