+++ DRS.pyproj
@@ -26,6 +26,9 @@
     <Compile Include="fddrescue.py" />
     <Compile Include="Fddrescue\ExcludeConvert.py" />
     <Compile Include="Fddrescue\fddrescue.py" />
+    <Compile Include="menu.py">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Include="rescue.py" />
     <Compile Include="Fddrescue\sync.py" />
     <Compile Include="Fddrescue\test.py" />
